requirements.txt
transformers>=4.44.0
sentence-transformers>=3.0.0
torch>=2.2.0
accelerate>=0.33.0
datasets>=2.19.0
scikit-learn>=1.3.0
pandas>=2.2.0
numpy>=1.26.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
tqdm>=4.66.0
jupyterlab>=4.0.0